% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db.R
\name{connect_db}
\alias{connect_db}
\title{connect_db}
\usage{
connect_db(username, local = FALSE)
}
\arguments{
\item{username}{A character string specifing the username.}

\item{local}{A boolean indicating if the connection is local or remote.}
}
\value{
a db connection
}
\description{
This function connect to the database
}
\examples{
\dontrun{
con <- connect_db(username = 'philippev',local=FALSE)
DBI::dbDisconnect(con)
}
}
