% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_table.R
\name{.load_db_mutiple_tbl}
\alias{.load_db_mutiple_tbl}
\alias{.load_db_multiple_tbl}
\title{Load multiple tables based on their names}
\usage{
.load_db_multiple_tbl(tbl_names, con)
}
\arguments{
\item{tbl_names}{A character vector of tbl_names to load. Tbl need to be on the database and a dedicated loading function.}

\item{con}{A connection to the database.}
}
\value{
A tibble of 2 columns. tbl_name and a list-col storing the table.
}
\description{
Internal function
}
\examples{
names <- c("tblSex", "tblWeights")
con <- connect_db(username = 'philippev',local=FALSE)
.load_db_multiple_tbl(tbl_names = names, con = con)
}
