% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_and_format_data.R
\name{filter_all_data}
\alias{filter_all_data}
\title{filter_all_data}
\usage{
filter_all_data(AllCall_Tidy, n_obs)
}
\arguments{
\item{AllCall_Tidy}{a df build with the function get_all_call_tidy}

\item{n_obs}{minimal number of observations}
}
\value{
a tibble
}
\description{
filter data to retain colonies with X observations
}
\examples{
con <- con <- DBI::dbConnect(RMySQL::MySQL(), user = 'philippev', password = getPass::getPass(),
dbname = 'Moleratdatabase', host = 'Kalahariresearch.org')
AllCall_Tidy <- get_all_call_tidy(con)
Elo_Data <- filter_all_data(AllCall_Tidy, 15)
}
