% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_and_format_data.R
\name{get_colony}
\alias{get_colony}
\title{get_colony}
\usage{
get_colony(AllCall, Membership)
}
\arguments{
\item{AllCall}{AllCall df}

\item{Membership}{Membership df}
}
\value{
a tibble
}
\description{
add the colony for the all call df
}
\examples{
#con <- DBI::dbConnect(RMySQL::MySQL(), user = 'philippev', password = getPass::getPass(),
#dbname = 'Moleratdatabase', host = 'Kalahariresearch.org')
#Membership <- get_membership(con)
#FocalCall <- get_focal_call(con)
#ScanCall <- get_scan_call(con)
#AllCall <- get_all_call(FocalCall, ScanCall)
#AllCall_colony <- get_colony(AllCall\%>\%
# select(Winner,Date)\%>\%
# rename(AnimalID=Winner),Membership)

}
