% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_and_format_data.R
\name{get_focal_calls}
\alias{get_focal_calls}
\alias{get_focal_call}
\title{get_focal_calls}
\usage{
get_focal_call(con)
}
\arguments{
\item{con}{a connection to the database}
}
\value{
a tibble
}
\description{
return focal call for all colonies
}
\examples{
#con <- DBI::dbConnect(RMySQL::MySQL(), user = 'philippev', password = getPass::getPass(),
#dbname = 'Moleratdatabase', host = 'Kalahariresearch.org')
#focal_call <- get_focal_call(con)

}
