% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_and_format_data.R
\name{get_membership}
\alias{get_membership}
\title{get_membership}
\usage{
get_membership(con)
}
\arguments{
\item{con}{a connection to the database}
}
\value{
a tibble
}
\description{
return the membership table from the database
}
\examples{
 #con <- DBI::dbConnect(RMySQL::MySQL(), user = 'philippev', password = getPass::getPass(),
 #dbname = 'Moleratdatabase', host = 'Kalahariresearch.org')
 #Membership <- get_membership(con)
}
