% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_and_format_data.R
\name{get_sex}
\alias{get_sex}
\title{get_sex}
\usage{
get_sex(con)
}
\arguments{
\item{con}{a connection to the database}
}
\value{
a tibble
}
\description{
return a tibble with the sex of the ID
}
\examples{
#con <- DBI::dbConnect(RMySQL::MySQL(), user = 'philippev', password = getPass::getPass(),
#dbname = 'Moleratdatabase', host = 'Kalahariresearch.org')
#sex_DF <- get_sex(con)
}
