% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_table.R
\name{create_db_tbl_family}
\alias{create_db_tbl_family}
\alias{create_db_tbl}
\alias{create_db_tbl_life_history}
\alias{create_db_tbl_behavior_scan}
\alias{create_db_tbl_behavior_focal}
\title{Create database tables}
\usage{
create_db_tbl_life_history(con)

create_db_tbl_behavior_scan(con)

create_db_tbl_behavior_focal(con)
}
\arguments{
\item{con}{A connection to the database}
}
\value{
This function returns a tibble
of 2 columns. tbl_name and a list-col storing the table.
}
\description{
The tables loaded by these functions will be stored in a hidden
environment. We do that so that other functions have access to it without the
user being tempted to interfere with it.
\itemize{
\item \link{create_db_tbl_life_history} will load the tables from the database relative to life_history
\item \link{create_db_tbl_behavior_scan} will load the tables from the database relative to scan behavior
\item \link{create_db_tbl_behavior_focal} will load the tables from the database relative to focal behavior
}
}
\details{
These functions can be used with a connection to the server of the Kalhari
research programme using the function \link{connect_db}.
}
\section{Functions}{
\itemize{
\item \code{create_db_tbl_life_history}: Create Life history db_tables

\item \code{create_db_tbl_behavior_scan}: create behavior db_tables for scan behavior

\item \code{create_db_tbl_behavior_focal}: create behavior db_tables for focal behavior
}}

\examples{

\dontrun{
load_db_tbl_life_history(con)
}
\dontrun{
create_db_tbl_behavior_scan(con)
}

\dontrun{
create_db_tbl_behavior_focal(con)
}

}
