% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db.R
\name{load_db_family}
\alias{load_db_family}
\alias{load_db_MembershipBetweenV2}
\alias{load_db_tblAnimalID}
\alias{load_db_tblSex}
\alias{load_db_tblWeights}
\alias{load_db_qry_BirthDate}
\alias{load_db_qry_DeathDate}
\alias{load_db_tblPairing}
\alias{load_db_tblColonyCodes}
\alias{load_db_tblLitterCode}
\alias{load_db_Escape}
\alias{load_db_Experiments_SubjectID}
\alias{load_db_PV_Parentage}
\alias{load_db_FirstColony}
\alias{load_db_ScanBehaviour_Instant_List}
\alias{load_db_ScanBehaviour_Cont_List}
\alias{load_db_tblScanSessionDetails}
\title{Load database tables}
\usage{
load_db_MembershipBetweenV2(con)

load_db_tblAnimalID(con)

load_db_tblSex(con)

load_db_tblWeights(con)

load_db_qry_BirthDate(con)

load_db_qry_DeathDate(con)

load_db_tblPairing(con)

load_db_tblColonyCodes(con)

load_db_tblLitterCode(con)

load_db_Escape(con)

load_db_Experiments_SubjectID(con)

load_db_PV_Parentage(con)

load_db_FirstColony(con)

load_db_ScanBehaviour_Instant_List(con)

load_db_ScanBehaviour_Cont_List(con)

load_db_tblScanSessionDetails(con)
}
\arguments{
\item{con}{A connection to the database}
}
\value{
This function returns nothing directly, but it creates a tibble of
tables and table names stored in a hidden environment.
}
\description{
\itemize{
\item load_db_XXX will load a specific table from the database.
}
}
\details{
These functions can be used with a connection to the server of the Kalhari
research programme using the function \link{connect_db}. For user level function
check \link{create_db_tbl_family}
}
\section{Functions}{
\itemize{
\item \code{load_db_MembershipBetweenV2}: load MembershipBetweenV2

\item \code{load_db_tblAnimalID}: load tblAnimalID

\item \code{load_db_tblSex}: load tblSex

\item \code{load_db_tblWeights}: load tblWeights

\item \code{load_db_qry_BirthDate}: load qry_BirthDate

\item \code{load_db_qry_DeathDate}: load qry_DeathDate

\item \code{load_db_tblPairing}: load tblPairing

\item \code{load_db_tblColonyCodes}: load tblColonyCodes

\item \code{load_db_tblLitterCode}: load tblLitterCode

\item \code{load_db_Escape}: load Escape

\item \code{load_db_Experiments_SubjectID}: load Experiments_SubjectID

\item \code{load_db_PV_Parentage}: load PV_Parentage

\item \code{load_db_FirstColony}: load FirstColony

\item \code{load_db_ScanBehaviour_Instant_List}: load ScanBehaviour_Instant_List

\item \code{load_db_ScanBehaviour_Cont_List}: load ScanBehaviour_Cont_List

\item \code{load_db_tblScanSessionDetails}: load tblScanSessionDetails
}}

\examples{

con <- connect_db(username = 'philippev',local=FALSE)
load_db_MembershipBetweenV2(con)
load_db_tblAnimalID(con)
load_db_tblSex(con)
load_db_tblWeights(con)
load_db_qry_BirthDate(con)
load_db_qry_DeathDate(con)
load_db_tblPairing(con)
load_db_tblColonyCodes(con)
load_db_tblLitterCode(con)
load_db_Escape(con)
load_db_Experiments_SubjectID(con)
load_db_PV_Parentage(con)
load_db_FirstColony(con)
load_db_ScanBehaviour_Instant_List(con)
load_db_ScanBehaviour_Cont_List(con)
load_db_tblScanSessionDetails(con)
}
